feat(navbar): Simplify navbar when logged out
