--- src/components/BootstrapMenu.svelte
@@ -1,199 +1,0 @@
-<script lang="ts">
-	import { _, locale } from "../i18n";
-	import { aboutPath, homePath, installPath, loginPath, securityPath } from "../router";
-	import { isLoginEnabled } from "../store";
-	import { link, useLocation } from "svelte-navigator";
-	import LanguageSelector from "./LanguageSelector.svelte";
-
-	interface Page {
-		path: string;
-		titleKey: `home.nav.${"home" | "about" | "security" | "install" | "log-in"}`;
-	}
-
-	let isNavButtonOpen = false; // This is exactly what Bootstrap does lol
-
-	const homeRoute = homePath();
-
-	function isNotNull<T>(tbd: T | null): tbd is T {
-		return tbd !== null;
-	}
-
-	const pages: Array<Page> = [
-		{ path: homePath(), titleKey: "home.nav.home" } as const,
-		{ path: aboutPath(), titleKey: "home.nav.about" } as const,
-		{ path: securityPath(), titleKey: "home.nav.security" } as const,
-		{ path: installPath(), titleKey: "home.nav.install" } as const,
-		isLoginEnabled ? ({ path: loginPath(), titleKey: "home.nav.log-in" } as const) : null,
-	].filter(isNotNull);
-
-	const location = useLocation();
-	$: currentPath = $location.pathname;
-
-	function toggle() {
-		isNavButtonOpen = !isNavButtonOpen;
-	}
-
-	function close() {
-		isNavButtonOpen = false;
-	}
-
-	$: $locale && close(); // Close the nav when locale changes
-</script>
-
-<nav class="navbar-147f68de navbar navbar-expand-md navbar-dark">
-	<a
-		href={homeRoute}
-		class="navbar-brand-9794720e"
-		role="heading"
-		aria-label={$_("common.platform")}
-		title={$_("common.platform")}
-		use:link>A&cent;countable</a
-	>
-	<button
-		class="navbar-toggler {!isNavButtonOpen ? 'collapsed' : ''}"
-		type="button"
-		data-bs-toggle="collapse"
-		data-bs-target="#navbarNav"
-		aria-controls="navbarNav"
-		aria-expanded={isNavButtonOpen}
-		aria-label={$_("app.toggle-nav")}
-		on:click={toggle}
-	>
-		<span class="navbar-toggler-icon" />
-	</button>
-
-	<div id="navbarNav" class="collapse navbar-collapse {isNavButtonOpen ? 'show' : ''}">
-		<ul class="navbar-nav">
-			{#each pages as page (page.path)}
-				<li class="nav-item {currentPath === page.path ? 'active' : ''}">
-					<a
-						class="nav-link {currentPath === page.path ? 'active' : ''}"
-						href={page.path}
-						on:click={close}
-						use:link
-						>{$_(page.titleKey)}
-						{#if currentPath === page.path}
-							<span class="visually-hidden">{$_("common.current-aside")}</span>
-						{/if}
-					</a>
-				</li>
-			{/each}
-			<li class="nav-item">
-				<div class="locale">
-					<LanguageSelector />
-				</div>
-			</li>
-		</ul>
-	</div>
-</nav>
-
-<style lang="scss">
-	@use "styles/setup" as *;
-	@use "styles/colors" as *;
-
-	// Import only the Bootstrap we need:
-	@import "bootstrap/scss/functions";
-	@import "bootstrap/scss/variables";
-	@import "bootstrap/scss/mixins";
-	@import "bootstrap/scss/reboot";
-	@import "bootstrap/scss/transitions";
-	@import "bootstrap/scss/nav";
-	@import "bootstrap/scss/navbar";
-	@import "bootstrap/scss/helpers/_visually-hidden.scss";
-
-	// $change: 575px; // Bootstrap's 'sm' change position
-	$change: 767px; // Bootstrap's 'md' change position
-
-	.navbar-brand-9794720e {
-		display: block;
-		font-weight: bold;
-		font-size: x-large;
-		z-index: 50;
-		margin-left: 16pt;
-		text-decoration: none;
-		border-radius: 4pt;
-		color: color($label-dark);
-
-		@media (hover: hover) {
-			&:hover {
-				color: color($link);
-				text-decoration: underline;
-			}
-		}
-
-		&:focus-visible {
-			outline: 2pt solid color($link);
-		}
-	}
-
-	nav.navbar.navbar-147f68de {
-		display: flex;
-		flex-flow: row nowrap;
-		width: 100%;
-
-		#navbarNav {
-			z-index: 100;
-
-			@include mq($until: $change) {
-				position: absolute;
-				top: 1.8em;
-				right: 0;
-				width: 100%;
-			}
-		}
-
-		ul,
-		.navbar-toggler {
-			list-style-type: none;
-			margin-left: auto;
-			margin-right: 8pt;
-			z-index: 50;
-		}
-
-		@include mq($until: $change) {
-			.navbar {
-				position: relative;
-
-				&-nav {
-					z-index: 100;
-					margin-top: 20pt;
-					padding: 0 8pt;
-					text-align: right;
-					min-width: 100vw;
-					background-color: color($navbar-background);
-				}
-			}
-		}
-
-		.active {
-			font-weight: bold;
-		}
-
-		.nav-item {
-			margin: auto 0;
-
-			> .locale {
-				width: fit-content;
-				margin-left: auto;
-			}
-
-			&.active {
-				color: color($link);
-			}
-
-			.nav-link {
-				border-radius: 4pt;
-
-				&:focus-visible {
-					outline: 2pt solid color($link);
-				}
-			}
-
-			@media (hover: hover) {
-				.nav-link:hover {
-					text-decoration: none;
-				}
-			}
-		}
-	}
-</style>
